May solve #20 (overflow in TM)
Added some Debug outputs (commented out). I'm really not sure if this
fixes the issue (but I think).

Note that you always can use this tool as random generator and pass the
generated wors to JFLAP.

An Example how to do this would be executing

`python3 main.py ... | grep True | head -n-1 | awk -F " " '{print $1}' > testTrue`
same with `False`.

The resulting file generated with this tool
can be imported in JFLAP via `Input -> Multiple Run -> Load Inputs`.

You can then check if all fields of the `True` file are filled with
`accepted` and all fields of the `False` file with `rejected` (load them
separately so you don't need to bother where one file starts and ends).

Of course this workflow is far from beeing pleasent, but at the moment
necessary to be sure the TM-Module makes the right descisions.

If you find any discrepancy please contact me for debugging (or make a
pull-request on your own of you want). If everything matches, you may
note this as well from time to time in #20 so that we can get a feeling
if its alright (maybe this workaround fixes the issue already).

The last paragraph of course only applies if the corresponding exercise
is not the active homework anymore ;)