Implemented an interface to the gotoLib
Note that this are the hints that are to be added to the `README.md`
after fully integrating this module into the `theoTool`

One should pay attention to the spacing of the elements of the
procedure. Every procedure element needs a label (they might have no
structure in common, so using `Rom:` and `M1:` along each other is
totally fine).

Currently following instructions are recognized (case-sensitive up to now):
- var := var/num
- var = var/num + var/num
- var = var/num - var/num
- var = var/num * var/num
- var = var/num div var/num
- var = var/num mod var/num
- if pred goto label
- goto label

Where pred might be one of the following:
- var/num = var/num
- var/num != var/num
- var/num >= var/num
- var/num <= var/num
- var/num < var/num
- var/num > var/num