Changed config principle!!!
The checkL function is now splitted up into multiple functions for each
module. The reasoning behind this is that regarding TMs as well as with
goto, there might not be a real true/false return value that has to be
checked.

With this change it is possible for the user to check not only the
true/false but the possible return value too.